Extract network/gateway orchestration into container_gateway.rs
Step 5 of splitting containers.rs. Moves ensure_network,
gateway_not_found_blocks, ensure_gateway, is_transient_recreate_error,
is_network_already_exists_error, refresh_gateway, remove_gateway, and
verify_environment_sites. refresh_gateway/remove_gateway/
verify_environment_sites are re-exported (6 external files depend on
them directly: site_commands.rs, livelink.rs, certificate_commands.rs);
ensure_gateway/ensure_network are only imported (operate_inner and
refresh_site_routes call them directly, no external consumers).

Narrower than originally planned: shutdown(), refresh_site_routes(),
and service_url() stay in containers.rs rather than moving too.
shutdown() iterates every environment (not gateway-specific) and only
touches the gateway as one part of app-exit cleanup; refresh_site_routes()
is built around prepare() (the core stack-generation orchestrator);
service_url() is built around the private service_context() helper
shared by several other CRUD functions. Moving any of the three would
have meant either loosening private-helper visibility for a single
external caller or creating a real two-way dependency between
containers.rs and container_gateway.rs beyond the one-directional
edge (gateway -> compose, for site_hostnames) already accounted for
in the plan — not worth it for concerns that aren't really "gateway".
1945 -> 1580 lines (container_gateway.rs: 385).
